+++ CHANGELOG.md
@@ -18,7 +18,9 @@
 
 ## New Functionality
 
-- [openApi-generator] Support the generation of transpiled JavaScript OpenAPI clients with the new flags: `generateJs` and `tsConfig`.
+- [openapi-generator] Support the generation of transpiled JavaScript OpenAPI clients with the new flags: `generateJs` and `tsConfig`.
+- [openapi-generator] Add options `writeReadme` for generating readme files.
+- [openapi-generator] Support copying files from given paths via a new option `additionalFiles`.
 
 ## Improvements