+++ CHANGELOG.md
@@ -18,7 +18,7 @@
 
 ## New Functionality
 
-- [openApi-generator] Support the generation of transpiled JavaScript OpenAPI clients with the new flags: `generateJs` and `tsConfig`.
+-
 
 ## Improvements
 
@@ -28,6 +28,20 @@
 
 -
 
+
+# 1.37.1
+
+Release Date: TBD<br>
+API Docs: https://sap.github.io/cloud-sdk/api/1.37.1<br>
+Blog: TBD<br>
+
+## New Functionality
+
+- [openApi-generator] Support the generation of transpiled JavaScript OpenAPI clients with the new flags: `generateJs` and `tsConfig`.
+
+## Fixed Issues
+
+- [generator] Allow enhanced generation with OpenAPI files for all valid OData specification file endings (including e. g. .xml and .EDMX).
 
 # 1.37.0